moving remaining iam plugins to test format
