+++ plugins/iam/usersMfaEnabled.js
@@ -1,117 +1,88 @@
 var async = require('async');
 var AWS = require('aws-sdk');
-
-function getPluginInfo() {
-	return {
-		title: 'Users MFA Enabled',
-		query: 'usersMfaEnabled',
-		category: 'IAM',
-		description: 'Ensures a multi-factor authentication device is enabled for all users within the account',
-		tests: {
-			usersMfaEnabled: {
-				title: 'Users MFA Enabled',
-				description: 'Ensures a multi-factor authentication device is enabled for all users within the account',
-				more_info: 'User accounts should have an MFA device setup to enable two-factor authentication',
-				link: 'http://docs.aws.amazon.com/IAM/latest/UserGuide/Using_ManagingPasswordPolicies.html',
-				recommended_action: 'Enable an MFA device for the user account',
-				results: []
-			}
-		}
-	};
-}
+var helpers = require('../../helpers');
 
 module.exports = {
-	title: getPluginInfo().title,
-	query: getPluginInfo().query,
-	category: getPluginInfo().category,
-	description: getPluginInfo().description,
-	more_info: getPluginInfo().more_info,
-	link: getPluginInfo().link,
-	tests: getPluginInfo().tests,
+	title: 'Users MFA Enabled',
+	category: 'IAM',
+	description: 'Ensures a multi-factor authentication device is enabled for all users within the account',
+	more_info: 'User accounts should have an MFA device setup to enable two-factor authentication',
+	link: 'http://docs.aws.amazon.com/IAM/latest/UserGuide/Using_ManagingPasswordPolicies.html',
+	recommended_action: 'Enable an MFA device for the user account',
 
 	run: function(AWSConfig, callback) {
-
+		var results = [];
+		
 		var iam = new AWS.IAM(AWSConfig);
-		var pluginInfo = getPluginInfo();
 
-		iam.listUsers({}, function(err, data){
-			if (err) {
-				pluginInfo.tests.usersMfaEnabled.results.push({
+		helpers.cache(iam, 'listUsers', function(err, data) {
+			if (err || !data || !data.Users) {
+				results.push({
 					status: 3,
 					message: 'Unable to query for user MFA status',
 					region: 'global'
 				});
 
-				return callback(null, pluginInfo);
+				return callback(null, results);
 			}
 
-			// Perform checks for establishing if MFA token is enabled
-			if (data && data.Users) {
-				if (data.Users.length) {
-					var good = [];
-					var bad = [];
+			if (!data.Users.length) {
+				results.push({
+					status: 0,
+					message: 'No user accounts found',
+					region: 'global'
+				});
+				
+				return callback(null, results);
+			}
 
-					async.eachLimit(data.Users, 20, function(user, cb){
-						if (!user.PasswordLastUsed) {
-							// Skip users without passwords since they won't be logging into the console
-							return cb();
-						}
-						iam.listMFADevices({UserName: user.UserName}, function(mfaErr, mfaData){
-							if (mfaErr) {
-								pluginInfo.tests.usersMfaEnabled.results.push({
-									status: 3,
-									message: 'Unable to query MFA device for user: ' + user.UserName,
+			var good = [];
+			var bad = [];
+
+			async.eachLimit(data.Users, 20, function(user, cb){
+				if (!user.PasswordLastUsed) {
+					// Skip users without passwords since they won't be logging into the console
+					return cb();
+				}
+				iam.listMFADevices({UserName: user.UserName}, function(mfaErr, mfaData){
+					if (mfaErr) {
+						results.push({
+							status: 3,
+							message: 'Unable to query MFA device for user: ' + user.UserName,
+							region: 'global',
+							resource: user.Arn
+						});
+					} else {
+						if (mfaData && mfaData.MFADevices) {
+							if (mfaData.MFADevices.length) {
+								results.push({
+									status: 0,
+									message: 'User: ' + user.UserName + ' has an MFA device',
 									region: 'global',
 									resource: user.Arn
 								});
 							} else {
-								if (mfaData && mfaData.MFADevices) {
-									if (mfaData.MFADevices.length) {
-										pluginInfo.tests.usersMfaEnabled.results.push({
-											status: 0,
-											message: 'User: ' + user.UserName + ' has an MFA device',
-											region: 'global',
-											resource: user.Arn
-										});
-									} else {
-										pluginInfo.tests.usersMfaEnabled.results.push({
-											status: 1,
-											message: 'User: ' + user.UserName + ' does not have an MFA device enabled',
-											region: 'global',
-											resource: user.Arn
-										});
-									}
-								}
+								results.push({
+									status: 1,
+									message: 'User: ' + user.UserName + ' does not have an MFA device enabled',
+									region: 'global',
+									resource: user.Arn
+								});
 							}
-							cb();
-						});
-					}, function(err){
-						if (err) {
-							pluginInfo.tests.usersMfaEnabled.results.push({
-								status: 3,
-								message: 'Unable to query for user MFA status',
-								region: 'global'
-							});
 						}
-						callback(null, pluginInfo);
-					});
-				} else {
-					pluginInfo.tests.usersMfaEnabled.results.push({
-						status: 0,
-						message: 'No user accounts found',
+					}
+					cb();
+				});
+			}, function(err){
+				if (err) {
+					results.push({
+						status: 3,
+						message: 'Unable to query for user MFA status',
 						region: 'global'
 					});
-					callback(null, pluginInfo);
 				}
-			} else {
-				pluginInfo.tests.usersMfaEnabled.results.push({
-					status: 3,
-					message: 'Unable to query for user MFA status',
-					region: 'global'
-				});
-
-				return callback(null, pluginInfo);
-			}
+				callback(null, results);
+			});
 		});
 	}
 };
